requirements.txt
opencv-python>=4.5.0
open3d>=0.15.0
numpy>=1.21.0
pyrealsense2>=2.50.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
matplotlib>=3.5.0
PyYAML>=6.0